New env variables for Strimzi and cert manager installation (#1003)
* added kroxylicious scale up and down test cases

* added two new Env variables for downstream pipeline

* reorder imports

* updated dev guide

* formatted

* using parseBoolean

* refactor isDeployed for certManager and renamed variable to skip strimzi installation

* looking for deployment instead of crd for cert manager

* suggested changes

* order imports

+++ kroxylicious-systemtests/src/main/java/io/kroxylicious/systemtests/installation/kroxylicious/CertManager.java
@@ -16,6 +16,7 @@ import io.fabric8.kubernetes.client.dsl.NamespaceListVisitFromServerGetDeleteRec
 
 import io.kroxylicious.systemtests.Constants;
 import io.kroxylicious.systemtests.utils.DeploymentUtils;
+import io.kroxylicious.systemtests.utils.NamespaceUtils;
 
 import static io.kroxylicious.systemtests.k8s.KubeClusterResource.kubeClient;
 
@@ -26,6 +27,7 @@ public class CertManager {
     private static final Logger LOGGER = LoggerFactory.getLogger(CertManager.class);
 
     private final NamespaceListVisitFromServerGetDeleteRecreateWaitApplicable<HasMetadata> deployment;
+    private boolean deleteCertManager = true;
 
     /**
      * Instantiates a new Cert manager.
@@ -37,13 +39,18 @@ public class CertManager {
                 .load(DeploymentUtils.getDeploymentFileFromURL(Constants.CERT_MANAGER_URL));
     }
 
+    private boolean isDeployed() {
+        return !kubeClient().getClient().apps().deployments().inAnyNamespace().withLabelSelector("app=cert-manager").list().getItems().isEmpty();
+    }
+
     /**
      * Deploy cert manager.
      */
     public void deploy() {
         LOGGER.info("Deploy cert manager in {} namespace", Constants.CERT_MANAGER_NAMESPACE);
-        if (kubeClient().getNamespace(Constants.CERT_MANAGER_NAMESPACE) != null) {
+        if (isDeployed()) {
             LOGGER.warn("Skipping cert manager deployment. It is already deployed!");
+            deleteCertManager = false;
             return;
         }
         deployment.create();
@@ -55,7 +62,12 @@ public class CertManager {
      * @throws IOException the io exception
      */
     public void delete() throws IOException {
+        if (!deleteCertManager) {
+            LOGGER.warn("Skipping cert manager deletion. It was previously installed");
+            return;
+        }
         LOGGER.info("Deleting Cert Manager in {} namespace", Constants.CERT_MANAGER_NAMESPACE);
         deployment.withGracePeriod(0).delete();
+        NamespaceUtils.deleteNamespaceWithWait(Constants.CERT_MANAGER_NAMESPACE);
     }
 }
